Reorganize the structure of our classes and moved several methods to the DOMupdates file
