+++ src/Game.js
@@ -1,13 +1,43 @@
 class Game {
-  constructor() {
+  constructor (data) {
+    this.categories = data.categories
+    this.cards = data.clues;
+    this.randomCategories = Object.entries(this.categories).sort((a, b) => 0.5 - Math.random())
+    this.roundCategories;
+    this.cardSet;
+    this.playerSet = [];
+  }
 
+  assignCategoriesRound1() {
+    this.roundCategories = this.randomCategories.slice(0, 4)
+    this.assignCards();
+  }
+
+  assignCategoriesRound2() {
+    this.roundCategories = this.randomCategories.slice(4, 9)
+    this.assignCards();
+  }
+
+  assignCategoriesRound3() {
+    this.roundCategories = this.randomCategories.slice(10)
+    this.assignCards();
+  }
+
+  assignCards() {
+    let pointValues = [100, 200, 300, 400]
+    let randomCardSet = []
+    let finalCardSet = []
+    let ids = this.roundCategories.map(el => el[1])
+    ids.forEach((id, index) => {
+      randomCardSet.push(this.cards.filter(el => el.categoryId === ids[index]))
+    })
+    pointValues.forEach((value, index1) => {
+      finalCardSet.push(randomCardSet.map(nestedArray => nestedArray.find(ely => ely.pointValue === pointValues[index1])))
+    })
+    this.cardSet = finalCardSet;
   }
 
   startGame() {
-  }
-
-  incrementRounds() {
-
   }
 
   determineWinner() {
