+++ src/Player.js
@@ -1,18 +1,15 @@
 class Player {
-  constructor(playerScore = 0, playerName, playerWager, turn, round) {
+  constructor(playerScore = 0, playerName, playerWager) {
     this.playerScore = playerScore;
     this.playerName = playerName;
     this.playerWager = playerWager;
     this.canWager = false;
-    this.turn = turn;
-    this.round = round;
   }
-
-  
 
   manageWagers() {
     if (this.playerScore > 5) {
       this.canWager = true;
+    }
   }
 }
