+++ src/index.js
@@ -1,0 +1,18 @@
+  //DOM.js event methods
+  // playerPickCategory() {
+  // }
+  // playerPickCardValue(valuePicked) {
+  // }
+  //combine with evaluate player guess
+  // changePlayerScore() {
+
+  // }
+  // evaluatePlayersGuesses(playerGuess, correctAnswer) {
+    
+  // }
+//   determineCurrentPlayer() {
+//   }
+
+//   determinePlayersEndOfTurn() {
+//   }
+// }
